+++ Tests/FunctionalTests/DependencyResolutionTests.swift
@@ -65,6 +65,15 @@ class DependencyResolutionTests: XCTestCase {
             XCTAssertEqual(output, "♣︎K\n♣︎Q\n♣︎J\n♣︎10\n♣︎9\n♣︎8\n♣︎7\n♣︎6\n♣︎5\n♣︎4\n")
         }
     }
+    
+    func testConvenienceBranchInit() throws {
+        fixture(name: "DependencyResolution/External/Branch") { prefix in
+            // Tests the convenience init .package(url: , branch: )
+            let app = prefix.appending(component: "Bar")
+            let result = try SwiftPMProduct.SwiftBuild.executeProcess([], packagePath: app)
+            XCTAssert(result.exitStatus == .terminated(code: 0))
+        }
+    }
 
     func testMirrors() {
         fixture(name: "DependencyResolution/External/Mirror") { prefix in